+++ NEWS.md
@@ -7,3 +7,7 @@
 * 0.0.1.900 modified setuphtml so that only the files listed in the resultTable file and the HTML and CSS files are deleted if cleanslate=TRUE. In this way any data files and other files stored in resdir remain safe.
 
 * 0.0.1.850 Modified the functions so that the output html files could have any desired name rather than having to accept the default of 'aMSE'. 
+
+* 0.0.1.800 Modified both addtable and addplot so that the input filename need not have the full path included.
+
+* 0.0.1.750 Further modificaiton to addplot, addtable, and setuphtml to simplify their use. Input filenames now must not include the path to resdir